requirements.1.txt
os 
logging 
dotenv 
binance.client 